+++ modules/services/views/src/main/java/au/com/shiftyjelly/pocketcasts/views/fragments/BaseFragment.kt
@@ -38,7 +38,7 @@ open class BaseFragment : Fragment(), CoroutineScope, HasBackstack {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         if (view.background == null) {
-            view.setBackgroundColor(view.context.getThemeColor(UR.attr.primary_ui_01))
+            view.setBackgroundColor(view.context.getThemeColor(UR.attr.primary_ui_04))
         }
         view.isClickable = true
         view.isFocusable = true
@@ -94,7 +94,6 @@ open class BaseFragment : Fragment(), CoroutineScope, HasBackstack {
     }
 
     override fun onPrepareOptionsMenu(menu: Menu) {
-        super.onPrepareOptionsMenu(menu)
         context?.let {
             menu.tintIcons(it.getThemeColor(UR.attr.secondary_icon_01))
         }